refactor: `S3FileUploader` 파일 위치를 service 패키지 내부로 이동
